+++ src/esm/VertTuple.js
@@ -10,7 +10,9 @@
  * @modified 2021-01-20 Added UID.
  * @modified 2022-02-02 Added the `destroy` method.
  * @modified 2023-09-29 Fixed a calculation error in the VertTuple.hasPoint() function; distance measure was broken!
- * @version 1.2.1
+ * @modified 2024-09-10 Chaging the first param of `pointDistance` from `Vertex` to less strict type `XYCoords`. This should not break anything.
+ * @modified 2024-09-10 Adding the optional `epsilon` param to the `hasPoint` method.
+ * @version 1.2.2
  */
 import { Vertex } from "./Vertex";
 import { UIDGenerator } from "./UIDGenerator";
@@ -217,21 +219,32 @@ export class VertTuple {
      * that point is located between point `a` and `b`.
      *
      * @method hasPoint
-     * @param {Vertex} point The point to check.
-     * @param {boolean=} insideBoundsOnly If set to to true (default=false) the point must be between start and end point of the line.
+     * @param {Vertex} point - The point to check.
+     * @param {boolean=} insideBoundsOnly - [optional] If set to to true (default=false) the point must be between start and end point of the line.
+     * @param {number=Vertex.EPSILON} epsilon - [optional] A tolerance.
      * @return {boolean} True if the given point is on this line.
      * @instance
      * @memberof VertTuple
      */
-    hasPoint(point, insideBoundsOnly) {
+    hasPoint(point, insideBoundsOnly, epsilon) {
         const t = this.getClosestT(point);
         // Compare to pointDistance?
         const distance = Math.sqrt(VertTuple.vtutils.dist2(point, this.vertAt(t)));
+        // console.log(
+        //   "distance",
+        //   distance,
+        //   "epsilon",
+        //   epsilon,
+        //   "distance < (epsilon ?? Vertex.EPSILON)",
+        //   distance < (epsilon ?? Vertex.EPSILON),
+        //   "distance < (epsilon ?? Vertex.EPSILON) && t >= 0 && t <= 1",
+        //   distance < (epsilon ?? Vertex.EPSILON) && t >= 0 && t <= 1
+        // );
         if (typeof insideBoundsOnly !== "undefined" && insideBoundsOnly) {
-            return distance < Vertex.EPSILON && t >= 0 && t <= 1;
+            return distance < (epsilon !== null && epsilon !== void 0 ? epsilon : Vertex.EPSILON) && t >= 0 && t <= 1;
         }
         else {
-            return distance < Vertex.EPSILON; // t >= 0 && t <= 1;
+            return distance < (epsilon !== null && epsilon !== void 0 ? epsilon : Vertex.EPSILON); // t >= 0 && t <= 1;
         }
     }
     /**
@@ -251,7 +264,7 @@ export class VertTuple {
      * The the minimal distance between this line and the specified point.
      *
      * @method pointDistance
-     * @param {Vertex} p The point (vertex) to measre the distance to.
+     * @param {XYCoords} p The point (vertex) to measre the distance to.
      * @return {number} The absolute minimal distance.
      * @instance
      * @memberof VertTuple